requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
nltk==3.8.1
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
qrcode[pil]==7.4.2
Pillow==10.0.1
requests==2.31.0
werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
tensorflow==2.15.0
